# Conflicts:
#	EmberApp/app/styles/app.scss

+++ src/UserBundle/Business/Manager/Agent/JsonApiUpdateAgentManagerTrait.php
@@ -35,6 +35,10 @@ trait JsonApiUpdateAgentManagerTrait
              */
             $agent = $data;
         }
+
+        if ($agent->getImage() && $agent->getImage()->getId() ===0){
+            $agent->getImage()->setId(null);
+        }
         /** @var Agent $dbAgent */
         $dbAgent       = $this->getEntityReference($agent->getId());
         $dbAgentGroup  = $dbAgent->getGroup();
@@ -89,16 +93,16 @@ trait JsonApiUpdateAgentManagerTrait
                 $this->eventDispatcher->dispatch(AgentEvents::ON_AGENT_GROUP_CHANGE, $event);
             }
 
-//            try {
-//                $syncResult = $this->syncWithTCRPortal($agent, 'edit');
-//                if (is_object($syncResult) && $syncResult->code == 200) {
+            try {
+                $syncResult = $this->syncWithTCRPortal($agent, 'edit');
+                if (is_object($syncResult) && $syncResult->code == 200) {
                     $this->flushDb();
-//                } else {
-//                    return new ArrayCollection(AgentApiResponse::AGENT_SYNC_ERROR_RESPONSE);
-//                }
-//            } catch (\Exception $exception) {
-//                return new ArrayCollection(AgentApiResponse::AGENT_SYNC_ERROR_RESPONSE);
-//            }
+                } else {
+                    return new ArrayCollection(AgentApiResponse::AGENT_SYNC_ERROR_RESPONSE);
+                }
+            } catch (\Exception $exception) {
+                return new ArrayCollection(AgentApiResponse::AGENT_SYNC_ERROR_RESPONSE);
+            }
         }
 
         return $this->createJsonAPiUpdateResponse($agentOrException);
